app/requirements.txt
openai==0.28.0
python-dotenv==0.21.1
typing==3.7.4.3